Move fontations binary execution functions into their own module.
