+++ ift/encoder/encoder.h
@@ -253,8 +253,9 @@ class Encoder {
                                          std::string url_template,
                                          common::CompatId compat_id);
 
-  absl::Status PopulateGlyphKeyedPatchMap(ift::proto::PatchMap& patch_map,
-                                          const design_space_t& design_space) const;
+  absl::Status PopulateGlyphKeyedPatchMap(
+      ift::proto::PatchMap& patch_map,
+      const design_space_t& design_space) const;
 
   absl::StatusOr<common::hb_face_unique_ptr> CutSubsetFaceBuilder(
       hb_face_t* font, const SubsetDefinition& def) const;
