+++ ift/encoder/encoder_test.cc
@@ -19,6 +19,7 @@
 #include "common/font_helper.h"
 #include "common/hb_set_unique_ptr.h"
 #include "gtest/gtest.h"
+#include "ift/client/fontations_client.h"
 #include "ift/proto/ift_table.h"
 #include "ift/proto/patch_map.h"
 
@@ -37,6 +38,7 @@ using common::FontData;
 using common::FontHelper;
 using common::hb_set_unique_ptr;
 using common::make_hb_set;
+using ift::client::ToGraph;
 using ift::proto::DEFAULT_ENCODING;
 using ift::proto::GLYPH_KEYED;
 using ift::proto::IFTTable;
@@ -92,17 +94,6 @@ class EncoderTest : public ::testing::Test {
     return result;
   }
 
-  Status to_file(const FontData& data, const char* path) {
-    FILE* f = fopen(path, "wb");
-    if (!f) {
-      return absl::InternalError("Unable to open file for output.");
-    }
-
-    fwrite(data.data(), 1, data.size(), f);
-    fclose(f);
-    return absl::OkStatus();
-  }
-
   btree_set<uint32_t> ToCodepointsSet(const FontData& font_data) {
     hb_face_t* face = font_data.reference_face();
 
@@ -141,79 +132,6 @@ class EncoderTest : public ::testing::Test {
     }
 
     return result;
-  }
-
-  absl::StatusOr<std::string> exec(const char* cmd) {
-    std::array<char, 128> buffer;
-    std::string result;
-    FILE* pipe = popen(cmd, "r");
-    if (!pipe) {
-      return absl::InternalError("Unable to start process.");
-    }
-    while (fgets(buffer.data(), static_cast<int>(buffer.size()), pipe) !=
-           nullptr) {
-      result += buffer.data();
-    }
-    if (pclose(pipe)) {
-      return absl::InternalError("exec command failed.");
-    }
-    return result;
-  }
-
-  void ParseGraph(const std::string& text, graph& out) {
-    std::stringstream ss(text);
-
-    std::string line;
-    while (getline(ss, line)) {
-      std::stringstream line_ss(line);
-      std::string node;
-      if (!getline(line_ss, node, ';')) {
-        continue;
-      }
-
-      auto& edges = out[node];
-
-      std::string edge;
-      while (getline(line_ss, edge, ';')) {
-        edges.insert(edge);
-      }
-    }
-  }
-
-  Status ToGraph(const Encoder& encoder, const FontData& base, graph& out) {
-    char template_str[] = "encoder_test_XXXXXX";
-    const char* temp_dir = mkdtemp(template_str);
-
-    if (!temp_dir) {
-      return absl::InternalError("Failed to create temp working directory.");
-    }
-
-    std::string font_path = absl::StrCat(temp_dir, "/font.ttf");
-    auto sc = to_file(base, font_path.c_str());
-    if (!sc.ok()) {
-      return sc;
-    }
-
-    for (auto& p : encoder.Patches()) {
-      auto& path = p.first;
-      auto& data = p.second;
-      std::string full_path = absl::StrCat(temp_dir, "/", path);
-      auto sc = to_file(data, full_path.c_str());
-      if (!sc.ok()) {
-        return sc;
-      }
-    }
-
-    std::string command =
-        absl::StrCat("${TEST_SRCDIR}/fontations/ift_graph --font=", font_path);
-    auto r = exec(command.c_str());
-    if (!r.ok()) {
-      return r.status();
-    }
-
-    ParseGraph(*r, out);
-
-    return absl::OkStatus();
   }
 };
